Move logging configuration to support package CIRC-362
