+++ src/main/java/org/folio/circulation/support/logging/Logging.java
@@ -1,10 +1,9 @@
-package org.folio.circulation;
+package org.folio.circulation.support.logging;
 
 public class Logging {
   private Logging() { }
 
-
-  static void initialiseFormat() {
+  public static void initialiseFormat() {
     System.setProperty("vertx.logger-delegate-factory-class-name",
       "io.vertx.core.logging.SLF4JLogDelegateFactory");
   }
